Fix integer overflow in large dataset databox allocation and indexing

+++ pftools/databox.h
@@ -65,7 +65,7 @@ extern "C" {
 typedef struct {
   double         *coeffs;
 
-  int nx, ny, nz;
+  unsigned int nx, ny, nz;
 
   /* other info that may not be available */
   double x, y, z;
@@ -111,7 +111,7 @@ typedef enum {
 
 /* databox.c */
 Databox *NewDatabox(int nx, int ny, int nz, double x, double y, double z, double dx, double dy, double dz);
-Databox *NewDataboxDefault(int nx, int ny, int nz, double x, double y, double z, double dx, double dy, double dz,
+Databox *NewDataboxDefault(size_t nx, size_t ny, size_t nz, double x, double y, double z, double dx, double dy, double dz,
                            double default_value);
 void GetDataboxGrid(Tcl_Interp *interp, Databox *databox);
 void SetDataboxGrid(Databox *databox, int nx, int ny, int nz, double x, double y, double z,
